修改 ISALIVE消息 的bug

+++ commmsg.h
@@ -22,7 +22,7 @@ private:
     /**
      * 文件下载进度
      */
-    qint8 rate;
+    double rate;
 
     /** 文件下载信息
      * downloadAddress 文件下载地址
@@ -45,7 +45,7 @@ private:
 public:
     CommMsg();
     CommMsg(TCPCtrlMsgType msgType);
-    CommMsg(TCPCtrlMsgType msgType, qint8 rate);
+    CommMsg(TCPCtrlMsgType msgType, double rate);
     CommMsg(TCPCtrlMsgType msgType, QString downloadAddress, int lenMax);
     CommMsg(TCPCtrlMsgType msgType, qint32 token, qint64 pos, qint32 len);
     CommMsg(TCPCtrlMsgType msgType, qint32 token, qint32 index);
